# docs/qa/openapi.yaml
openapi: 3.0.3
info:
  title: DocNear API
  version: 1.0.0
  description: Shared patient, doctor, owner, admin and Telegram API. JSON responses
    wrap payloads in success/data; see docs/api-contract.md.
paths:
...
  /api/appointments/my/:
    get:
      operationId: appointments_my_retrieve
      tags:
      - appointments
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Appointment'
          description: ''
...
  /api/doctors/{id}/availability/:
    get:
      operationId: doctors_availability_retrieve
      parameters:
      - in: query
        name: clinic_id
        schema:
          type: integer
          minimum: 1
        required: true
      - in: query
        name: date
        schema:
          type: string
          format: date
        required: true
      - in: path
        name: id
        schema:
          type: integer
        description: A unique integer value identifying this doctor profile.
        required: true
      tags:
      - doctors
      security:
      - jwtAuth: []
      - {}
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Availability'
          description: ''
...
  /api/telegram/appointments/my/:
    get:
      operationId: telegram_appointments_my_retrieve
      tags:
      - telegram
      security:
      - TelegramBotSecret: []
        TelegramLinkedUser: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Appointment'
          description: ''
...
  /api/telegram/doctors/{id}/availability/:
    get:
      operationId: telegram_doctors_availability_retrieve
      parameters:
      - in: query
        name: clinic_id
        schema:
          type: integer
          minimum: 1
        required: true
      - in: query
        name: date
        schema:
          type: string
          format: date
        required: true
      - in: path
        name: id
        schema:
          type: integer
        required: true
      tags:
      - telegram
      security:
      - jwtAuth: []
      - {}
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Availability'
          description: ''
...
components:
  schemas:
...
    Affiliation:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
        clinic:
          type: integer
        clinic_name:
          type: string
          readOnly: true
        specialty:
          type: integer
        specialty_name:
          type: string
          readOnly: true
        latitude:
          type: number
          format: double
          readOnly: true
        longitude:
          type: number
          format: double
          readOnly: true
        is_primary:
          type: boolean
        consultation_duration:
          type: integer
          maximum: 240
          minimum: 5
        buffer_time:
          type: integer
          maximum: 120
          minimum: 0
        max_appointments_per_day:
          type: integer
          maximum: 32767
          minimum: 1
        is_active:
          type: boolean
      required:
      - clinic
      - clinic_name
      - id
      - latitude
      - longitude
      - specialty
      - specialty_name
...
    Appointment:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
        booking_id:
          type: string
          readOnly: true
        patient:
          type: integer
          readOnly: true
        patient_name:
          type: string
          readOnly: true
        doctor:
          type: integer
          readOnly: true
        doctor_name:
          type: string
          readOnly: true
        clinic:
          type: integer
          readOnly: true
        clinic_name:
          type: string
          readOnly: true
        specialty:
          type: integer
          readOnly: true
        specialty_name:
          type: string
          readOnly: true
        appointment_date:
          type: string
          format: date
          readOnly: true
        start_time:
          type: string
          format: time
          readOnly: true
        end_time:
          type: string
          format: time
          readOnly: true
        patient_note:
          type: string
          readOnly: true
        status:
          allOf:
          - $ref: '#/components/schemas/StatusEnum'
          readOnly: true
        cancel_reason:
          type: string
          readOnly: true
        created_at:
          type: string
          format: date-time
          readOnly: true
        updated_at:
          type: string
          format: date-time
          readOnly: true
      required:
      - appointment_date
      - booking_id
      - cancel_reason
      - clinic
      - clinic_name
      - created_at
      - doctor
      - doctor_name
      - end_time
      - id
      - patient
      - patient_name
      - patient_note
      - specialty
      - specialty_name
      - start_time
      - status
      - updated_at
...
    Availability:
      type: object
      properties:
        date:
          type: string
          format: date
        doctor:
          $ref: '#/components/schemas/Doctor'
        clinic:
          type: object
          additionalProperties: {}
        slots:
          type: array
          items:
            $ref: '#/components/schemas/Slot'
      required:
      - clinic
      - date
      - doctor
      - slots
...
    Doctor:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
        name:
          type: string
          readOnly: true
        affiliations:
          type: array
          items:
            $ref: '#/components/schemas/Affiliation'
          readOnly: true
        distance_km:
          type: number
          format: double
          readOnly: true
        next_available_time:
          type: string
          readOnly: true
        created_at:
          type: string
          format: date-time
          readOnly: true
        updated_at:
          type: string
          format: date-time
          readOnly: true
        profile_image:
          type: string
          format: uri
          nullable: true
        bio:
          type: string
        experience_years:
          type: integer
          maximum: 32767
          minimum: 0
        education:
          type: string
        certifications: {}
        languages: {}
        rating:
          type: string
          format: decimal
          pattern: ^-?\d{0,1}(?:\.\d{0,2})?$
          readOnly: true
        total_reviews:
          type: integer
          readOnly: true
        is_verified:
          type: boolean
        is_active:
          type: boolean
        accepts_bookings:
          type: boolean
      required:
      - affiliations
      - created_at
      - distance_km
      - id
      - name
      - next_available_time
      - rating
      - total_reviews
      - updated_at
...
    Slot:
      type: object
      properties:
        time:
          type: string
        end_time:
          type: string
        available:
          type: boolean
      required:
      - available
      - end_time
      - time
...
    StatusEnum:
      enum:
      - pending
      - confirmed
      - waiting
      - in_progress
      - completed
      - cancelled
      - rejected
      - no_show
      type: string
      description: |-
        * `pending` - Pending
        * `confirmed` - Confirmed
        * `waiting` - Waiting
        * `in_progress` - In progress
        * `completed` - Completed
        * `cancelled` - Cancelled
        * `rejected` - Rejected
        * `no_show` - No show
...
  securitySchemes:
    TelegramBotSecret:
      type: apiKey
      in: header
      name: X-Telegram-Bot-Secret
      description: Server-side bot secret; never ship in a patient client.
    TelegramLinkedUser:
      type: apiKey
      in: header
      name: X-Telegram-User-Id
      description: Telegram ID of the patient linked by a one-time code.
    jwtAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT